Merge the current working out of order DTMF support to a branch
So, this gets things relatively close.  res_rtp_asterisk has basic
knowledge of whether or not a jitter buffer exists that can hold
DTMF frames, and stops doing 'frame dropping' if that happens.

Unfortunately, the jitter buffer doesn't handle two things very well:
1) Collisions, which occurs when DTMF END events happen (usually 3
in a row)
2) Multiple frames in a list are inheriting the timestamp of the first,
so that doesn't work terribly well
3) DTMF frames are getting interpolated, and that's just no good
at all.  If a DTMF frame doesn't exist, we need to switch to a different
jitter buffer for that cycle.
4) We may want to not change the timer frequency.  The length of a DTMF
is fairly long, during which period we may be putting voice frames
that need to get read.  I might move this back towards just keeping the
timer at a minimum frequency.

I'm trying to prevent putting a whole bunch of frame type checking
into func_jitterbuffer, but it looks like there's going to have to
be some level of that.

So, still some stuff to do.  Need to think about some of the collision
issues.