Check in work to avoid race condition when querying config objects
Instead of storing and swapping out the config objects indivdually and
possibly getting a reload in between a request for a privates container
and the private config info container, store all of the config objects
in a single object that we swap out and get back. This object represents
a snapshot of all config data so that everything is 100% consistent.