Make aco_process_category_options and aco_set_defaults take an aco_info
This was a hold-over from before the aco_info struct existed. The options
container is private to the aco_info, so if these are to be public functions
they need to take an aco_info and not a container. I still think there could
be situations where it is useful to have these public, so I'm leaving them
here for now. I may revisit this decision.