Take a lazier approach to adding missing message IDs to voicemails.
Rather than perusing every single voicemail message on startup, we
now do it more-or-less on demand. Whenever a voicemail message
snapshot is created, we will ensure that the message has a message
ID associated with it.

IMAP is still not addressed here because it is a whole other can
of worms and will greatly detract from the clarity of the diff that
will be posted to reviewboard.