Add support for updating a message ID when using IMAP storage.
IMAP messages, once delivered, cannot be updated. Instead, the
method used here is to delete the message and then store it again
with the message ID added on.

In my next update, I'm going to see if I can preserve the "read" vs.
"unread" status of the message when storing it so that messages don't
appear mistakenly as unread.