Add new message-ID-having message to proper IMAP folder.
IMAP has the drawback, at the moment anyway, that new
messages are created and automatically saved in the INBOX instead
of using the directory name to try to determine the folder. So what
has to be done instead is to save the message, then move it to
the appropriate folder.

I adjusted save_to_folder() to have a new parameter that indicates
whether the message should be copied or moved. I also added
documentation to indicate that this final parameter is only
observed when using IMAP voicemail.