I hate chan_agent.
Given the tools Asterisk has now, there are some potentially
neat things we could be doing with agent-style channels and there
are most definitely better ways to implement the channel type.

I'm experimenting here with some stuff. I don't really have a
solid plan, nor do I expect this to actually end up in Asterisk
as-is. This is a sandbox, a playground if you will.