Some initial tinkering with agent channels.
This employs a state machine for moving agents to
their various states throughout the process. Currently
two states are defined. Moving beyond this will require
me to start defining the channel tech for agents, which
is where the trickery will begin.

If this ever makes it in, a lot of this will probably
be split into multiple files since the actual channel
part of agents will be minimal compared to the state
machine and other logic associated with them.