Progress towards agent channel.
This is where the fun happens. When an agent logs
in, we actually override/decorate certain channel
tech callbacks with our own versions. Currently, I
have the requester and call methods defined. There
will need to be a hangup method defined and most
likely a read, as well. Those will wait until a
separate time though.

Next step on the agenda will be to devise the
configuration code so that we can start configuring
agents.