Change login operation.
Now instead of finding an agent to use, we allocate a new one
on login. This means that there is no longer a need for the
logged out state, so it has been removed.

I have this fear that RAII_VAR may be my hammer and I'm 
seeing everything as nails right now...